requirements.txt
hydra-core==1.1.1
matplotlib==3.4.3
numpy==1.19.0
omegaconf==2.1.0
pandas==1.3.4
pytorch_lightning==1.5.4
torch==1.10.0
tqdm==4.62.3
